\name{bic.mixture.multivariate}
\alias{bic.mixture.multivariate}
\title{Description: Latent class analysis based on (infinite) Gaussian mixture model. If the input (dat) is data matrix, a multivariate model is fitted.}
\usage{
  bic.mixture.multivariate(x, max.modes, ...)
}
\arguments{
  \item{x}{matrix (for multivariate analysis)}

  \item{max.modes}{Maximum number of modes to be checked
  for mixture model selection}

  \item{...}{Further optional arguments to be passed

  Returns:}
}
\value{
  Fitted latent class model (parameters and free energy)
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("netresponse")
}
\keyword{utilities}

