% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{centerData}
\alias{centerData}
\title{Center data matrix.}
\usage{
centerData(X, rm.na = TRUE, meanvalue = NULL)
}
\arguments{
\item{X}{The data set: samples x features. Each feature will be centered.}

\item{rm.na}{Ignore NAs.}

\item{meanvalue}{Can be used to set a desired center value. The default is
0.}
}
\value{
Centered data matrix.
}
\description{
Center data matrix to 0 for each variable by removing the means.
}
\note{
Note that the model assumes samples x features matrix, and centers
each feature.
}
\examples{

#centerData(X)

}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse").
}
\keyword{maths}
\keyword{utilities}

