\name{check.matrix}
\alias{check.matrix}
\title{check.matrix}
\description{Mostl for internal purposes. Check input matrix format.}
\usage{
check.matrix(datamatrix)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datamatrix}{See detect.responses}
}
\value{The datamatrix, possibly added with necessary formatting for the netresponse algorithm.}
\references{
See citation("netresponse")
}
\author{Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao.
        Maintainer: Leo Lahti \email{leo.lahti@iki.fi}}
\seealso{detect.responses}
\examples{
# datamatrix <- check.matrix(datamatrix)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }

