% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetResponseModel-accessors.R
\docType{methods}
\name{get.dat,NetResponseModel-method}
\alias{get.dat}
\alias{get.dat,NetResponseModel-method}
\title{Get subnetwork data}
\usage{
\S4method{get.dat}{NetResponseModel}(model, subnet.id, sample = NULL)
}
\arguments{
\item{model}{Result from NetResponse (detect.responses function).}

\item{subnet.id}{Subnet identifier. A natural number which specifies one of
the subnetworks within the 'model' object.}

\item{sample}{Define the retrieved samples}
}
\value{
Subnet data matrix
}
\description{
Get subnetwork data
}
\examples{
## Load a pre-calculated netresponse model obtained with 
# model <- detect.responses(toydata$emat, toydata$netw, verbose = FALSE)
# data( toydata ); get.dat(toydata$model) 
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{utilities}

