\name{get.model}
\alias{get.model}
\title{get.model}
\usage{
  get.model(datamatrix, network, edge, network.nodes, G,
    params)
}
\arguments{
  \item{datamatrix}{datamatrix}

  \item{network}{network}

  \item{edge}{edge}

  \item{network.nodes}{network.nodes}

  \item{G}{G}

  \item{params}{parameters}
}
\value{
  NetResponse model object
}
\description{
  Mainly for internal use. Calculate joint model for a
  given node pair.
}
\examples{
# TBA
}
\author{
  Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao.
  Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("netresponse")
}
\keyword{internal}

