\name{ICMg.get.comp.memberships}
\alias{ICMg.get.comp.memberships}
\title{ ICMg.get.comp.memberships }
\description{ Internal function for ICMg tools. }
\usage{ ICMg.get.comp.memberships(links, samples) }
\author{ Juuso Parkkinen }

\references{
Parkkinen, J. and Kaski, S. Searching for functional gene modules with interaction component models.
BMC Systems Biology 4 (2010), 4.
}

\keyword{ internal }
