% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list.responses.continuous.R
\name{list.responses.continuous.multi}
\alias{list.responses.continuous.multi}
\title{Description: Investigate association of a continuous variable and the modes; given a list of groupings}
\usage{
list.responses.continuous.multi(annotation.df, groupings, method = "t-test",
  pth = Inf, verbose = TRUE, rounding = NULL)
}
\arguments{
\item{annotation.df}{annotation data.frame with discrete factor levels, rows
named by the samples}

\item{groupings}{Sample mode information. Each element corresponds to one grouping; each grouping lists samples for the modes within that grouping.}

\item{method}{method for quantifying the association}

\item{pth}{p-value threshold applied to adjusted p-values}

\item{verbose}{verbose}

\item{rounding}{rounding digits

Returns:}
}
\value{
Table listing all associations between the factor levels and responses
}
\description{
Arguments:
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{utilities}

