\name{mixture.model}
\alias{mixture.model}
\title{Description: Fit Gaussian mixture model}
\usage{
  mixture.model(mat, vars, params, ...)
}
\arguments{
  \item{mat}{data matrix (for multivariate analysis)}

  \item{vars}{features to investigate in the data matrix}

  \item{params}{parameters}

  \item{...}{Further optional arguments to be passed

  Returns:}
}
\value{
  List with two elements: model: fitted mixture model
  (parameters and free energy); model.params: model
  parameters
}
\description{
  Arguments:
}
\author{
  Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("netresponse")
}
\keyword{utilities}

