<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> master
% Please edit documentation in R/mixture.densities.R
\name{P.rs.joint}
\alias{P.rs.joint}
\title{Description: Joint probabiity density for mode and sample group
Mainly for internal use; documentation will be provided later. Tools for calculating densities with Gaussian mixture models.}
\usage{
P.rs.joint(dat, pars, log = TRUE)
}
<<<<<<< HEAD
\arguments{
\item{dat}{features x samples data matrix for mixture modeling}

\item{pars}{Gaussian mixture model parameters (diagonal covariances); list with elements mu (mean vectors), sd (covariance diagonals), w (weights). The mu and sd are component x features matrices, w is vector giving weight for each component.}

\item{log}{Logical. Return densities in log domain.

Returns:}
}
\value{
Probability density
}
\description{
Arguments:
=======
\description{
Arguments:
 @param dat features x samples data matrix for mixture modeling
 @param pars Gaussian mixture model parameters (diagonal covariances); list with elements mu (mean vectors), sd (covariance diagonals), w (weights). The mu and sd are component x features matrices, w is vector giving weight for each component.
 @param log Logical. Return densities in log domain.
}
\details{
Returns:
  @return Probability density
>>>>>>> master
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}
\keyword{utilities}

