% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.densities.R
\name{P.s.r}
\alias{P.s.r}
\title{Description: Probabiity density for sample given mode
Mainly for internal use; documentation will be provided later. Tools for calculating densities with Gaussian mixture models.}
\usage{
P.s.r(dat, pars, log = TRUE)
}
\description{
Arguments:
 @param dat features x samples data matrix for mixture modeling
 @param pars Gaussian mixture model parameters (diagonal covariances); list with elements mu (mean vectors), sd (covariance diagonals), w (weights). The mu and sd are component x features matrices, w is vector giving weight for each component.
 @param log Logical. Return densities in log domain.
}
\details{
Returns:
  @return Probability density
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}
\keyword{utilities}

