\name{pick.model.parameters}
\alias{pick.model.parameters}
\title{pick.model.parameters}
\usage{
  pick.model.parameters(m, nodes)
}
\arguments{
  \item{m}{vdp.mixt output}

  \item{nodes}{node names for naming purposes}
}
\value{
  Model parameters
}
\description{
  pick.model.parameters
}
\examples{
#
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("netresponse")
}
\keyword{internal}

