% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.response}
\alias{plot.response}
\title{plot.response}
\usage{
\method{plot}{response}(x, mynet, mybreaks, mypalette, plot.names =
TRUE, colors = TRUE, plot.type = "twopi", ...)
}
\arguments{
\item{x}{A numerical vector, or NULL.}

\item{mynet}{Binary matrix specifying the interactions between nodes.}

\item{mybreaks}{Specify breakpoints for color plot.}

\item{mypalette}{Specify palette for color plot.}

\item{plot.names}{Plot node names (TRUE) or indices (FALSE).}

\item{colors}{Plot colors. Logical.}

\item{plot.type}{Network plot mode. For instance, 'neato' or 'twopi'.}

\item{...}{Further arguments for plot function.}
}
\value{
Used for its side-effects.
}
\description{
Plot a specific transcriptional response for a given subnetwork.
}
\examples{
#tmp <- plot.response(model, mynet,
#  	maintext = paste("Subnetwork", subnet.id))
}
\author{
Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao. Maintainer:
Leo Lahti <leo.lahti@iki.fi>
}
\references{
L. Lahti et al.: Global modeling of transcriptional responses in
interaction networks. Submitted.
}
\keyword{utilities}

