<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> master
% Please edit documentation in R/plot-methods.R
\name{plot_associations}
\alias{plot_associations}
\title{Association strength between category labels and responses.}
\usage{
plot_associations(x, subnet.id, labels, method = "hypergeometric",
  mode = "group.by.classes", ...)
}
\arguments{
\item{x}{NetResponseModel object}

\item{subnet.id}{Subnetwork.}

\item{labels}{Factor. Labels for the data samples. Name by samples, or
provide in the same order as in the original data.}

\item{method}{Method to calculate association strength.}

\item{mode}{group.by.responses or group.by.classes: indicate barplot
grouping type.}

<<<<<<< HEAD
\item{...}{Other arguments to be passed for plot.}
=======
\item{...}{Other arguments to be passed for plot_}
>>>>>>> master
}
\value{
Used for side effect (plotting).
}
\description{
<<<<<<< HEAD
Plot association strength between user-defined category labels and
responses in a selected subnetwork.
}
\details{
Associations are showm in terms -log10(p) enrichment values for the
annotation categories for the responses within the specified subnetwork. No
correction for multiple testing.
=======
Plot association strength between user-defined category labels and responses in a selected subnetwork. Associations are showm in terms -log10(p) enrichment values for the annotation categories for the responses within the specified subnetwork. No correction for multiple testing.
>>>>>>> master
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse").
}
\seealso{
<<<<<<< HEAD
plot.responses
=======
plot_responses
>>>>>>> master
}
\keyword{utilities}

