% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_data}
\alias{plot_data}
\title{Plot observed data.}
\usage{
plot_data(x, subnet.id, labels, ...)
}
\arguments{
\item{x}{NetResponseModel object.}

\item{subnet.id}{Specify the subnetwork.}

\item{labels}{Annotation categories.}

\item{...}{Further arguments for plot function.}
}
\value{
ggplot2 plot object
}
\description{
Plotting tool for measurement data.
Produces boxplot for each feature in each annotation category for the
selected subnetwork.
}
\examples{
#
}
\author{
Leo Lahti <leo.lahti@iki.fi>
}
\references{
See citation("netresponse")
}
\seealso{
plot_responses
}
\keyword{utilities}

