% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_response}
\alias{plot_response}
\title{plot_response}
\usage{
plot_response(x, mynet, mybreaks, mypalette, plot_names = TRUE,
  colors = TRUE, plot_type = "twopi", ...)
}
\arguments{
\item{x}{A numerical vector, or NULL.}

\item{mynet}{Binary matrix specifying the interactions between nodes.}

\item{mybreaks}{Specify breakpoints for color plot_}

\item{mypalette}{Specify palette for color plot_}

\item{plot_names}{Plot node names (TRUE) or indices (FALSE).}

\item{colors}{Plot colors. Logical.}

\item{plot_type}{Network plot mode. For instance, 'neato' or 'twopi'.}

\item{...}{Further arguments for plot function.}
}
\value{
Used for its side-effects.
}
\description{
Plot a specific transcriptional response for a given subnetwork.
TRUE, colors = TRUE, plot_type = "twopi", ...)
}
\examples{
#tmp <- plot_response(model, mynet, 
#  	maintext = paste("Subnetwork", subnet.id))
}
\author{
Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao. Maintainer:
Leo Lahti <leo.lahti@iki.fi>
}
\references{
L. Lahti et al.: Global modeling of transcriptional responses in
interaction networks. Submitted.
}
\keyword{utilities}

