% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_subnet}
\alias{plot_subnet}
\title{plot_subnet}
\usage{
plot_subnet(x, subnet.id, network, plot_names = TRUE, ...)
}
\arguments{
\item{x}{Result from NetResponse (detect.responses function).}

\item{subnet.id}{Subnet id.}

\item{network}{Original network used in the modelling.}

\item{plot_names}{Plot node names (TRUE) or indices (FALSE).}

\item{...}{Further arguments for plot function.}
}
\value{
Used for its side-effects. Returns a matrix that describes the
investigated subnetwork.
}
\description{
Plot the given subnetwork.
}
\examples{
#
# res <- detect.responses(D, netw, verbose = FALSE)
# net <- plot_subnet(res, subnet.idx = 1)
}
\author{
Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao. Maintainer:
Leo Lahti <leo.lahti@iki.fi>
}
\references{
L. Lahti et al.: Global modeling of transcriptional responses in
interaction networks. Submitted.
}
\keyword{utilities}

