% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{PlotMixtureBivariate}
\alias{PlotMixtureBivariate}
\title{PlotMixtureBivariate}
\usage{
PlotMixtureBivariate(x, means, sds, ws, labels = NULL, confidence = 0.95,
  main = "", ...)
}
\arguments{
\item{x}{data matrix (samples x features)}

\item{means}{mode centroids (modes x features)}

\item{sds}{mode standard deviations, assuming diagonal covariance 
matrices (modes x features, each row giving the sqrt of covariance 
diagonal for the corresponding mode)}

\item{ws}{weight for each mode}

\item{labels}{Optional: sample class labels to be indicated in colors.}

\item{confidence}{Confidence interval for the responses based on the
covariances of each response. If NULL, no plotting.}

\item{main}{title text}

\item{...}{Further arguments for plot function.}
}
\value{
Used for its side-effects.
}
\description{
Visualize data, centroids and response confidence intervals for a given
Gaussian mixture model in two-dimensional (bivariate) case. Optionally, 
color the samples according to annotations labels.
}
\examples{
#plotMixture(dat, means, sds, ws)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse") for citation details.
}
\keyword{utilities}

