% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(x, subnet.id, labels = NULL, confidence = 0.95, ...)
}
\arguments{
\item{x}{NetResponseModel object. Output from the detect.responses function.}

\item{subnet.id}{Subnetwork id. Either character as 'Subnetwork-2' or
numeric as 2, which is then converted to character.}

\item{labels}{Optional: sample class labels to be indicated in colors.}

\item{confidence}{Confidence interval for the responses based on the
covariances of each response. If NULL, no plotting.}

\item{...}{Further arguments for plot function.}
}
\value{
Used for its side-effects.
}
\description{
Visualize data, centroids and response confidence intervals for a given
subnetwork with PCA. Optionally, color the samples according to annotations
labels.
}
\examples{
#plotPCA(x, subnet.id)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse") for citation details.
}
\keyword{utilities}

