% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{set.breaks}
\alias{set.breaks}
\title{Set breaks}
\usage{
set.breaks(mat, interval = 0.1)
}
\arguments{
\item{mat}{Matrix to visualize.}

\item{interval}{Density of color breakpoints.}
}
\value{
A vector listing the color breakpoints.
}
\description{
Set breakpoints for two-way color palette.
}
\examples{
 set.breaks(array(rnorm(100), dim = c(10, 10)), interval = .1) 
}
\author{
Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao. Maintainer:
Leo Lahti <leo.lahti@iki.fi>
}
\references{
L. Lahti et al.: Global modeling of transcriptional responses in
interaction networks. Submitted.
}
\keyword{utilities}

