\name{write.netresponse.results}
\alias{write.netresponse.results}
\title{Write NetResponse results summary into a file.}
\description{Write NetResponse results summary into a file.}
\usage{
write.netresponse.results(x, subnet.ids = NULL, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{NetResponseModel}
  \item{subnet.ids}{List of subnet ids to consider. By default, all subnets.}
  \item{filename}{Output file name.}
}
\details{Experimental version.}
\value{Used for side effects.}
\references{See citation("netresponse")}
\author{Leo Lahti \email{leo.lahti@iki.fi}}
\examples{
#
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }

