<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mixtures.R
\name{bic.mixture.univariate}
\alias{bic.mixture.univariate}
\title{Description: Latent class analysis based on (infinite) Gaussian mixture
model. If the input (dat) is data matrix, a multivariate model is fitted. If
the input is a vector or a 1-dimensional matrix, a univariate model is
fitted.}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.mixture.univariate.R
\name{bic.mixture.univariate}
\alias{bic.mixture.univariate}
\title{Univariate BIC mixture}
>>>>>>> master
\usage{
bic.mixture.univariate(x, max.modes, bic.threshold = 0, min.modes = 1, ...)
}
\arguments{
\item{x}{dat vector (for univariate analysis) or a matrix (for multivariate analysis)}

\item{max.modes}{Maximum number of modes to be checked for mixture model selection}

\item{bic.threshold}{BIC threshold which needs to be exceeded before a new mode is added to the mixture.}

\item{min.modes}{minimum number of modes}

<<<<<<< HEAD
\item{...}{Further optional arguments to be passed

Returns:}
=======
\item{...}{Further optional arguments to be passed}
>>>>>>> master
}
\value{
Fitted latent class model (parameters and free energy)
}
\description{
<<<<<<< HEAD
Arguments:
=======
Latent class analysis based on (infinite) Gaussian mixture
model. If the input (dat) is data matrix, a multivariate model is fitted. If
the input is a vector or a 1-dimensional matrix, a univariate model is
fitted.
>>>>>>> master
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{utilities}

