% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.mixture.univariate.R
\name{bic.mixture.univariate}
\alias{bic.mixture.univariate}
\title{Univariate BIC mixture}
\usage{
bic.mixture.univariate(x, max.modes, bic.threshold = 0, min.modes = 1, ...)
}
\arguments{
\item{x}{dat vector (for univariate analysis) or a matrix (for multivariate analysis)}

\item{max.modes}{Maximum number of modes to be checked for mixture model selection}

\item{bic.threshold}{BIC threshold which needs to be exceeded before a new mode is added to the mixture.}

\item{min.modes}{minimum number of modes}

\item{...}{Further optional arguments to be passed}
}
\value{
Fitted latent class model (parameters and free energy)
}
\description{
Latent class analysis based on (infinite) Gaussian mixture
model. If the input (dat) is data matrix, a multivariate model is fitted. If
the input is a vector or a 1-dimensional matrix, a univariate model is
fitted.
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{utilities}

