% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.select.best.mode.R
\name{bic.select.best.mode}
\alias{bic.select.best.mode}
\title{Select best mode with BIC}
\usage{
bic.select.best.mode(x, max.modes, bic.threshold, min.modes = 1)
}
\arguments{
\item{x}{dat vector (for univariate analysis) or a matrix (for multivariate analysis)}

\item{max.modes}{Maximum number of modes to be checked for mixture model selection}

\item{bic.threshold}{BIC threshold which needs to be exceeded before a new mode is added to the mixture.}

\item{min.modes}{Optiomal. Minimum number of modes.}
}
\value{
Fitted latent class model (parameters and free energy)
}
\description{
Select optimal number of mixture components by adding components until 
the increase in objective function is below threshold.
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{utilities}

