\name{check.matrix}
\alias{check.matrix}
\title{check.matrix}
\usage{
  check.matrix(datamatrix)
}
\arguments{
  \item{datamatrix}{See detect.responses}
}
\value{
  The datamatrix, possibly added with necessary formatting
  for the netresponse algorithm.
}
\description{
  Mostl for internal purposes. Check input matrix format.
}
\examples{
# datamatrix <- check.matrix(datamatrix)
}
\author{
  Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao.
  Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
  See citation("netresponse")
}
\seealso{
  detect.responses
}
\keyword{internal}

