% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.network.R
\name{check.network}
\alias{check.network}
\title{check.network}
\usage{
check.network(network, datamatrix, verbose = FALSE)
}
\arguments{
\item{network}{Input network, see detect.responses}

\item{datamatrix}{Input datamatrix, see detect.responses}

\item{verbose}{Print intermediate messages}
}
\value{
\item{formatted }{Formatted network (self-links removed)}
  \item{original }{Original network (possible in another representation format)} 
  \item{delta }{Cost function changes corresponding to the 'formatted' network.} 
  \item{nodes }{Nodes corresponding to the 'formatted' network.}
}
\description{
Internal use to check input network and format detect.responses.
}
\examples{
# check.network(network, datamatrix, verbose = FALSE)
}
\references{
See citation('netresponse')
}
\seealso{
detect.responses
}
\author{
Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
