% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.responses.R
\name{factor.responses}
\alias{factor.responses}
\title{Factor responses}
\usage{
factor.responses(annotation.vector, groupings, method = "hypergeometric",
  min.size = 2, data = NULL)
}
\arguments{
\item{annotation.vector}{annotation vector with discrete factor levels, and named by the samples}

\item{groupings}{List of groupings. Each model should have a sample-cluster assignment matrix qofz, or a vector of cluster indices named by the samples.}

\item{method}{method for enrichment calculation}

\item{min.size}{minimum sample size for a response}

\item{data}{data (samples x features; or a vector in univariate case)}
}
\value{
List with each element corresponding to one factor level and listing the responses according to association strength
}
\description{
List responses for each level of the given factor
}
\author{
Contact: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{utilities}

