% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.netw.R
\name{filter.netw}
\alias{filter.netw}
\title{filter.netw}
\usage{
filter.netw(network, delta, datamatrix, params)
}
\arguments{
\item{network}{network}

\item{delta}{associated cost function value changes for each node merge}

\item{datamatrix}{datamatrix}

\item{params}{parameters}
}
\value{
Filtered network
}
\description{
Mostly for internal use. Prefilter edges if speedups required.
}
\details{
Include only edges with the highest mutual information,
calculated based on the first principal components.
}
\examples{
#
}
\author{
Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao.  Maintainer:
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}

