% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mis.R
\name{get.mis}
\alias{get.mis}
\title{get.mis}
\usage{
get.mis(datamatrix, network, delta, network.nodes, G, params)
}
\arguments{
\item{datamatrix}{datamatrix}

\item{network}{network}

\item{delta}{delta}

\item{network.nodes}{network.nodes}

\item{G}{G}

\item{params}{params}
}
\value{
mutual information matrix
}
\description{
Mainly for internal use. Estimate mutual information for node pairs based on
the first principal components
}
\examples{
#
}
\author{
Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}

