\name{get.model}
\alias{get.model}
\title{get.model}
\description{Mainly for internal use. Calculate joint model for a given node pair.}
\usage{
get.model(datamatrix, network, edge, network.nodes, G, params)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datamatrix}{datamatrix}
  \item{network}{network}
  \item{edge}{edge}
  \item{network.nodes}{network.nodes}
  \item{G}{G}
  \item{params}{parameters}
}

\value{
NetResponse model object
}

\references{
See citation("netresponse")
}
\author{Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao.
        Maintainer: Leo Lahti \email{leo.lahti@iki.fi}}
\examples{
# TBA
}
\keyword{ internal }
