% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICMg.R
\name{ICMg.get.comp.memberships}
\alias{ICMg.get.comp.memberships}
\title{ICMg.get.comp.memberships}
\usage{
ICMg.get.comp.memberships(links, samples)
}
\arguments{
\item{links}{N x 2 matrix of link endpoints (N = number of links).}

\item{samples}{Posterior samples, as given by either ICMg.combined.sampler
or ICMg.links.sampler.}
}
\value{
A matrix containing the component memberships for each data point
(node).
}
\description{
Function for computing the component memberships for each data point from
the MCMC samples.
}
\author{
Juuso Parkkinen
}
\references{
Parkkinen, J. and Kaski, S. Searching for functional gene
modules with interaction component models. BMC Systems Biology 4 (2010), 4.
}
\seealso{
\code{\link{ICMg.combined.sampler}},
\code{\link{ICMg.links.sampler}}
}
\keyword{methods}

