\name{independent.models}
\alias{independent.models}
\title{independent.models}
\description{Mainly for internal use. Provide independent models for each node.}
\usage{independent.models(datamatrix, params)}
\arguments{
  \item{datamatrix}{datamatrix}
  \item{params}{parameters}
}
\value{
  \item{nodes }{Model for each node}
  \item{C }{Costs for individual models}

}
\references{
See citation("netresponse")
}
\author{Leo Lahti \email{leo.lahti@iki.fi}}

\examples{
# TBA
}
\keyword{ internal }

