% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent.models.R
\name{independent.models}
\alias{independent.models}
\title{independent.models}
\usage{
independent.models(datamatrix, params)
}
\arguments{
\item{datamatrix}{datamatrix}

\item{params}{parameters}
}
\value{
\item{nodes }{Model for each node} 
  \item{C }{Costs for individual models}
}
\description{
Mainly for internal use. Provide independent models for each node.
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}

