\docType{class}
\name{NetResponseModel-class}
\alias{get.dat,NetResponseModel-method}
\alias{NetResponseModel-class}
\alias{[[,NetResponseModel-method}
\alias{show,NetResponseModel-method}
\title{Class "NetResponseModel"}
\description{
  A NetResponse model.
}
\section{Objects from the Class}{
  Returned by \code{\link{detect.responses}} function.
}
\examples{
showClass("NetResponseModel")
}
\author{
  Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{classes}

