% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.model.pairs.R
\name{pick.model.pairs}
\alias{pick.model.pairs}
\title{Pick model pairs}
\usage{
pick.model.pairs(network, network.nodes, node.models, datamatrix, params)
}
\arguments{
\item{network}{network}

\item{network.nodes}{network.nodes}

\item{node.models}{node.models}

\item{datamatrix}{datamatrix}

\item{params}{parameters}
}
\value{
\item{model.pairs}{joint models for each node pair}
\item{delta}{corresponding delta value for the cost function}
}
\description{
Mainly for internal use. Calculate joint model for each node pair
}
\examples{
#
}
\author{
Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}

