<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internals.R
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.model.parameters.R
>>>>>>> master
\name{pick.model.parameters}
\alias{pick.model.parameters}
\title{pick.model.parameters}
\usage{
pick.model.parameters(m, nodes)
}
\arguments{
\item{m}{vdp.mixt output}

\item{nodes}{node names for naming purposes}
}
\value{
Model parameters
}
\description{
<<<<<<< HEAD
pick.model.parameters
=======
Pick model parameters
>>>>>>> master
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}

