% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.expression}
\alias{plot.expression}
\title{plot.expression}
\usage{
\method{plot}{expression}(x, maintext, ...)
}
\arguments{
\item{x}{samples x features matrix}

\item{maintext}{main title}

\item{...}{optional arguments}
}
\value{
Used for its side effects.
}
\description{
Plot expression matrix in color scale. For one-channel data; plot expression
of each gene relative to its mean expression level over all samples. Blue
indicates decreased expression and red indicates increased expression.
Brightness of the color indicates magnitude of the change. Black denotes no
change.
}
\examples{
#

#plot.expression(x)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse").
}
\seealso{
\code{\link{plot.scale}}
}
\keyword{utilities}

