% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMixture.R
\name{PlotMixture}
\alias{PlotMixture}
\title{Plot mixtures}
\usage{
PlotMixture(x, qofz, binwidth = 0.05, xlab.text = NULL, ylab.text = NULL,
  title.text = NULL)
}
\arguments{
\item{x}{data vector}

\item{qofz}{Mode assignment probabilities for each sample. Samples x modes.}

\item{binwidth}{binwidth for histogram}

\item{xlab.text}{xlab.text}

\item{ylab.text}{ylab.text}

\item{title.text}{title.text}
}
\value{
Used for its side-effects
}
\description{
Plot mixtures.
}
\examples{
# PlotMixture(x, qofz)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse") for citation details.
}
\keyword{utilities}

