% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.netw.R
\name{remove.negative.edges}
\alias{remove.negative.edges}
\title{remove.negative.edges}
\usage{
remove.negative.edges(network, delta, datamatrix)
}
\arguments{
\item{network}{network}

\item{delta}{associated cost function value changes for each node merge}

\item{datamatrix}{datamatrix}
}
\value{
Filtered network
}
\description{
Remove edges where the connected nodes correlate negatively
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}

