% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.enrichment.R
\name{response.enrichment}
\alias{response.enrichment}
\title{Enrichment for a specified sample group in the given response.}
\usage{
response.enrichment(total.samples, response.samples, annotated.samples,
  method = "hypergeometric")
}
\arguments{
\item{total.samples}{All samples in the data}

\item{response.samples}{Samples in the investigated subset}

\item{annotated.samples}{Samples at the investigated annotation level for enrichment calculation}

\item{method}{Enrichment method.}
}
\value{
List with enrichment statistics, depending on enrichment method.
}
\description{
Calculate enrichment values for a specified sample group in the given
response.
}
\examples{
#enr <- response.enrichment(subnet.id, models, sample, response, method)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\seealso{
order.responses
}
\keyword{utilities}

