% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.model.pair.R
\name{update.model.pair}
\alias{update.model.pair}
\title{update.model.pair}
\usage{
update.model.pair(datamatrix, delta, network, edge, network.nodes, G,
params, model.nodes, model.pairs)
}
\arguments{
\item{datamatrix}{datamatrix}

\item{delta}{delta}

\item{network}{network}

\item{edge}{edge}

\item{network.nodes}{network.nodes}

\item{G}{G}

\item{params}{params}

\item{model.nodes}{model.nodes}

\item{model.pairs}{model.pairs

Return:}
}
\value{
\item{model.pairs }{model.pairs} \item{delta }{delta}
}
\description{
Mainly for internal use. Calculate joint model for given node pair and
update delta accordingly.
}
\examples{
#
}
\author{
Leo Lahti, Olli-Pekka Huovilainen and Antonio Gusmao.  Maintainer:
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{internal}

