% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write.netresponse.results}
\alias{write.netresponse.results}
\title{Write NetResponse results summary into a file.}
\usage{
write.netresponse.results(x, subnet.ids = NULL, filename)
}
\arguments{
\item{x}{NetResponseModel}

\item{subnet.ids}{List of subnet ids to consider. By default, all subnets.}

\item{filename}{Output file name.}
}
\value{
Used for side effects.
}
\description{
Write NetResponse results summary into a file.
}
\details{
Experimental version.
}
\examples{
#
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("netresponse")
}
\keyword{utilities}

